Ángel Hernández/DJANGO/modelo349/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-environ
django-filter
django-cors-headers
